Use AppSyncRealTimeClient 1.0.2 Interceptors
