+++ AWSAppSyncClient/Internal/SubscriptionFactory/APIKeyBasedConnectionPool.swift
@@ -19,21 +19,12 @@ class APIKeyBasedConnectionPool: SubscriptionConnectionPool {
 
     func connection(for url: URL, connectionType: SubscriptionConnectionType) -> SubscriptionConnection {
 
-        let connectionProvider = endPointToProvider[url.absoluteString] ?? createConnectionProvider(for: url, connectionType: connectionType)
+        let connectionProvider = endPointToProvider[url.absoluteString] ??
+            ConnectionProviderFactory.createConnectionProvider(for: url,
+                                                               authInterceptor: APIKeyAuthInterceptor(apiKeyProvider.getAPIKey()),
+                                                               connectionType: connectionType)
         endPointToProvider[url.absoluteString] = connectionProvider
         let connection = AppSyncSubscriptionConnection(provider: connectionProvider)
         return connection
-    }
-
-    func createConnectionProvider(for url: URL, connectionType: SubscriptionConnectionType) -> ConnectionProvider {
-        let provider =  ConnectionPoolFactory.createConnectionProvider(for: url, connectionType: connectionType)
-        if let messageInterceptable = provider as? MessageInterceptable {
-            messageInterceptable.addInterceptor(APIKeyAuthInterceptor(apiKeyProvider))
-        }
-        if let connectionInterceptable = provider as? ConnectionInterceptable {
-            connectionInterceptable.addInterceptor(RealtimeGatewayURLInterceptor())
-            connectionInterceptable.addInterceptor(APIKeyAuthInterceptor(apiKeyProvider))
-        }
-        return provider
     }
 }
