+++ AWSAppSyncClient/Internal/SubscriptionFactory/IAMBasedConnectionPool.swift
@@ -21,23 +21,14 @@ class IAMBasedConnectionPool: SubscriptionConnectionPool {
 
     func connection(for url: URL, connectionType: SubscriptionConnectionType) -> SubscriptionConnection {
 
-        let connectionProvider = endPointToProvider[url.absoluteString] ?? createConnectionProvider(for: url, connectionType: connectionType)
+        let connectionProvider = endPointToProvider[url.absoluteString] ??
+            ConnectionProviderFactory.createConnectionProvider(for: url,
+                                                               authInterceptor: IAMAuthInterceptor(credentialProvider,
+                                                                                                   region: regionType),
+                                                               connectionType: connectionType)
+
         endPointToProvider[url.absoluteString] = connectionProvider
         let connection = AppSyncSubscriptionConnection(provider: connectionProvider)
         return connection
     }
-
-    func createConnectionProvider(for url: URL, connectionType: SubscriptionConnectionType) -> ConnectionProvider {
-        let provider = ConnectionPoolFactory.createConnectionProvider(for: url, connectionType: connectionType)
-        if let messageInterceptable = provider as? MessageInterceptable {
-            messageInterceptable.addInterceptor(IAMAuthInterceptor(credentialProvider, region: regionType))
-        }
-        if let connectionInterceptable = provider as? ConnectionInterceptable {
-            connectionInterceptable.addInterceptor(RealtimeGatewayURLInterceptor())
-            connectionInterceptable.addInterceptor(IAMAuthInterceptor(credentialProvider, region: regionType))
-        }
-
-        return provider
-    }
-
 }
